+++ src/polymer/create-element-extension.ts
@@ -21,24 +21,11 @@ import {Rule} from '../rule';
 import {getDocumentContaining, stripIndentation, stripWhitespace} from '../util';
 
 
-class DomCallsToNative extends Rule {
-  code = 'dom-calls-to-native';
+class CreateElementExtension extends Rule {
+  code = 'create-element-extension';
   description = stripIndentation(`
-      Warns when the Polymer.dom is used in places where native
-      methods can now be used instead.
-
-        Polymer.dom(event).path
-
-      Accepted syntax:
-
-        event.composedPath()
+    Warns when using the second parameter of createElement for element extension using the is attribute.
   `);
-
-  private _replacements: Map<string, string> = new Map([
-    ['localTarget', 'target'],
-    ['rootTarget', 'composedPath()[0]'],
-    ['path', 'composedPath()']
-  ]);
 
   async check(document: Document) {
     const warnings: Warning[] = [];
@@ -48,55 +35,58 @@ class DomCallsToNative extends Rule {
     for (const doc of docs) {
       babelTraverse(doc.parsedDocument.ast, {
         noScope: true,
-        MemberExpression: (path) => {
-          if (!babel.isIdentifier(path.node.property)) {
-            return;
-          }
-
-          const name = path.node.property.name;
-          const replacement = this._replacements.get(name);
-
-          if (!this._isPolymerDomCall(path.node.object) ||
-              replacement === undefined) {
+        CallExpression: (path) => {
+          if (!this.isExtendingElementCall(path.node)) {
             return;
           }
 
           const containingDoc =
               getDocumentContaining(doc.sourceRange, document);
-
-          if (!containingDoc) {
+          if (containingDoc === undefined) {
             return;
           }
 
-          const sourceRange = containingDoc.sourceRangeForNode(path.node);
+          const sourceRange =
+              containingDoc.sourceRangeForNode(path.node.arguments[1]);
           if (sourceRange === undefined) {
             return;
           }
 
-          warnings.push(new Warning({
-            parsedDocument: document.parsedDocument,
-            code: 'deprecated-dom-call',
-            severity: Severity.WARNING, sourceRange,
-            message: stripWhitespace(`
-              Polymer.dom no longer needs to be called for "${name}",
-              instead "event.${replacement}" may be used.
-            `)
-          }));
-        }
+          if (babel.isStringLiteral(path.node.arguments[1])) {
+            warnings.push(new Warning({
+              parsedDocument: document.parsedDocument,
+              code: 'create-element-string-extension',
+              severity: Severity.WARNING, sourceRange,
+              message: stripWhitespace(`
+                document.createElement with a custom element tagname as the second parameter is deprecated.
+              `)
+            }));
+          } else {
+            warnings.push(new Warning({
+              parsedDocument: document.parsedDocument,
+              code: 'create-element-is-property-extension',
+              severity: Severity.WARNING, sourceRange,
+              message: stripWhitespace(`
+                Creating a custom element extension of a built-in element is not widely supported, and is not recommended.
+              `)
+            }));
+          }
+        },
       });
     }
 
     return warnings;
   }
 
-  private _isPolymerDomCall(expr: babel.Expression): boolean {
+  private isExtendingElementCall(expr: babel.Expression): boolean {
     return babel.isCallExpression(expr) &&
         babel.isMemberExpression(expr.callee) &&
         babel.isIdentifier(expr.callee.object) &&
+        expr.callee.object.name === 'document' &&
         babel.isIdentifier(expr.callee.property) &&
-        expr.callee.object.name === 'Polymer' &&
-        expr.callee.property.name === 'dom';
+        expr.callee.property.name === 'createElement' &&
+        expr.arguments.length === 2;
   }
 }
 
-registry.register(new DomCallsToNative());
+registry.register(new CreateElementExtension());