+++ src/test/polymer/create-element-extension_test.ts
@@ -21,7 +21,7 @@ import {registry} from '../../registry';
 import {WarningPrettyPrinter} from '../util';
 
 const fixtures_dir = path.join(__dirname, '..', '..', '..', 'test');
-const ruleId = 'dom-calls-to-native';
+const ruleId = 'create-element-extension';
 
 suite(ruleId, () => {
   let analyzer: Analyzer;
@@ -49,20 +49,16 @@ suite(ruleId, () => {
     const {warnings} = await linter.lint([`${ruleId}/${ruleId}.html`]);
     assert.deepEqual(warningPrinter.prettyPrint(warnings), [
       `
-      Polymer.dom(event).localTarget.innerHTML = 'foo';
-      ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~`,
+      document.createElement('style', 'custom-style');
+                                      ~~~~~~~~~~~~~~`,
       `
-      const path = Polymer.dom(event).path;
-                   ~~~~~~~~~~~~~~~~~~~~~~~`,
-      `
-      const root = Polymer.dom(event).rootTarget;
-                   ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~`,
+      document.createElement('ul', {is: 'x-expanding-list'});
+                                   ~~~~~~~~~~~~~~~~~~~~~~~~`,
     ]);
 
     assert.deepEqual(warnings.map((w) => w.message), [
-      'Polymer.dom no longer needs to be called for "localTarget", instead "event.target" may be used.',
-      'Polymer.dom no longer needs to be called for "path", instead "event.composedPath()" may be used.',
-      'Polymer.dom no longer needs to be called for "rootTarget", instead "event.composedPath()[0]" may be used.'
+      'document.createElement with a custom element tagname as the second parameter is deprecated.',
+      'Creating a custom element extension of a built-in element is not widely supported, and is not recommended.',
     ]);
   });
 });